+++ data_utils_split.py
@@ -48,22 +48,30 @@ nlp = spacy.load("en_core_web_sm")
 #     return embedding_matrix
 
 def build_wordvec_matrix(word2idx, embed_dim, type):
-    embedding_matrix_file_name = '{0}_{1}_embedding_matrix.pkl'.format(str(embed_dim), type)
-    print('loading word vectors ...')
-    embedding_matrix = np.zeros((len(word2idx), embed_dim))
-    embedding_matrix[1, :] = np.random.uniform(-1/np.sqrt(embed_dim), 1/np.sqrt(embed_dim), (1, embed_dim))
-    fname = 'snli_w2v'
-    snli_w2v = Word2Vec.load(fname)
-    word_vec = snli_w2v.wv
-    print('building embedding_matrix:', embedding_matrix_file_name)
-    for word, i in word2idx.items():
-        try:
-            vec = word_vec.get_vector(word, norm=True)
-        except:
-            continue
-        if vec is not None:
-            embedding_matrix[i] = vec
-    pickle.dump(embedding_matrix, open(embedding_matrix_file_name, 'wb'))
+
+    embedding_matrix_file_name = '{0}_{1}_{2}_embedding_matrix.pkl'.format(str(embed_dim), type, str(config['train_split']))
+
+    if os.path.exists(embedding_matrix_file_name):
+        print('loading embedding_matrix:', embedding_matrix_file_name)
+        embedding_matrix = pickle.load(open(embedding_matrix_file_name, 'rb'))
+
+    else:
+        print('loading word vectors ...')
+        embedding_matrix = np.zeros((len(word2idx), embed_dim))
+        embedding_matrix[1, :] = np.random.uniform(-1/np.sqrt(embed_dim), 1/np.sqrt(embed_dim), (1, embed_dim))
+        fname = 'snli_w2v'
+        snli_w2v = Word2Vec.load(fname)
+        word_vec = snli_w2v.wv
+        print('building embedding_matrix:', embedding_matrix_file_name)
+        for word, i in word2idx.items():
+            try:
+                vec = word_vec.get_vector(word, norm=True)
+            except:
+                continue
+            if vec is not None:
+                embedding_matrix[i] = vec
+        pickle.dump(embedding_matrix, open(embedding_matrix_file_name, 'wb'))
+        print("Saved embedding matrix\n")
 
     return embedding_matrix
 
@@ -126,18 +134,37 @@ class ABSADataset(object):
 class ABSADatasetReader:
     @staticmethod
     def __read_text__(fnames):
-        max_len = 0 # SOS and EOS
-        text = ''
-        for fname in fnames:    
-            for line in fname:
-                line = line.lower().strip()
-                line = re.sub(r"([.!?])", r" \1", line)
-                line = re.sub(r"[^a-zA-Z.!?]+", r" ", line)
-                temp = nlp(line)
-                temp = [item.text for item in temp]
-                if len(temp) > max_len:
-                    max_len = len(temp)
-                text += line + " "
+
+        if os.path.exists("raw_text_" + str(config['train_split'])):
+            with open("raw_text_" + str(config['train_split']), 'rb') as f:
+                text = pickle.load(f)
+            f.close()
+            with open("max_len_" + str(config['train_split']), 'rb') as f:
+                max_len = pickle.load(f)
+            f.close()
+        
+        else:
+
+            max_len = 0 # SOS and EOS
+            text = ''
+            for fname in fnames:    
+                for line in fname:
+                    line = line.lower().strip()
+                    line = re.sub(r"([.!?])", r" \1", line)
+                    line = re.sub(r"[^a-zA-Z.!?]+", r" ", line)
+                    temp = nlp(line)
+                    temp = [item.text for item in temp]
+                    if len(temp) > max_len:
+                        max_len = len(temp)
+                    text += line + " "
+
+            with open("raw_text_" + str(config['train_split']), "wb") as f:
+                pickle.dump(text, f)
+            f.close()
+            with open("max_len_" + str(config['train_split']), "wb") as m:
+                pickle.dump(max_len, m)
+            m.close()
+            
         return text, max_len
 
     @staticmethod
@@ -159,42 +186,62 @@ class ABSADatasetReader:
         #     fin = open('./snli_test_nonsvo.graph', 'rb')
         #     nonsvo_graph = pickle.load(fin)
         #     fin.close()
+        if os.path.exists(flag + "_data_" + str(config['train_split'])) and flag == 'train':
+            print("Loading in from pickle\n")
+            with open(flag + "_data_" + str(config['train_split']), 'rb') as f:
+                all_data = pickle.load(f)
+            f.close()
+            with open(flag + "_total_" + str(config['train_split']), 'rb') as f:
+                text_total = pickle.load(f)
+            f.close()
 
-        all_data = []
-        text_total = []
-        # graph_id = 0
-        for line in fname:
-            
-            line = line.lower().strip()
-            line = re.sub(r"([.!?])", r" \1", line)
-            line = re.sub(r"[^a-zA-Z.!?]+", r" ", line)
-            line = nlp(line)
-            listified = [item.text for item in line]
-            # line = line.split(' ')
-            text_total.append(listified)
+        else:
 
-            text_indices = tokenizer.text_to_sequence(listified)
-            text_padding = [config['PAD_token']] * (max_len - len(text_indices))
-            text_indices = [config['SOS_token']] + text_indices + [config['EOS_token']] + text_padding
-            # svo = svo_graph[graph_id]
-            # nonsvo = nonsvo_graph[graph_id]
-            # print(svo)
-            # print(nonsvo)
-            # print(line)
-            # print(flag)
-            # exit()
+            all_data = []
+            text_total = []
+            # graph_id = 0
+            for line in fname:
+                
+                line = line.lower().strip()
+                line = re.sub(r"([.!?])", r" \1", line)
+                line = re.sub(r"[^a-zA-Z.!?]+", r" ", line)
+                line = nlp(line)
+                listified = [item.text for item in line]
+                # line = line.split(' ')
+                if flag == 'val' or flag == 'test':
+                    text_total.append(listified)
 
-            data = {
-                'context': line,
-                'text_indices': text_indices,
-            }
+                text_indices = tokenizer.text_to_sequence(listified)
+                text_padding = [config['PAD_token']] * (max_len - len(text_indices))
+                text_indices = [config['SOS_token']] + text_indices + [config['EOS_token']] + text_padding
+                # svo = svo_graph[graph_id]
+                # nonsvo = nonsvo_graph[graph_id]
+                # print(svo)
+                # print(nonsvo)
+                # print(line)
+                # print(flag)
+                # exit()
 
-            all_data.append(data)
+                data = {
+                    'context': line,
+                    'text_indices': text_indices,
+                }
 
-        return all_data, text_total
+                all_data.append(data)
+           
+            with open(flag + "_data_" + str(config['train_split']), "wb") as f:
+                pickle.dump(all_data, f)
+            with open(flag + "_total_" + str(config['train_split']), "wb") as f:
+                pickle.dump(text_total, f)
+
+        if flag == 'val' or flag == 'test':
+            return all_data, text_total
+        else:
+            return all_data
 
     def __init__(self, dataset, train, val, test, split, embed_dim=300):
         text, max_len = ABSADatasetReader.__read_text__([train, test])
+        print(len(train))
         self.max_len = max_len
 
         # if os.path.exists(dataset+'_gcn_word2idx.pkl'):
@@ -210,7 +257,8 @@ class ABSADatasetReader:
         #     pickle.dump(self.tokenizer.idx2word, f)
         # self.embedding_matrix = build_embedding_matrix(self.tokenizer.word2idx, embed_dim, dataset)
         self.embedding_matrix = build_wordvec_matrix(self.tokenizer.word2idx, embed_dim, dataset)
-        self.train_data, self.text_train = ABSADataset(ABSADatasetReader.__read_data__(train, self.tokenizer, max_len, flag = 'train'))
+        print(self.embedding_matrix.shape)
+        # self.train_data = ABSADataset(ABSADatasetReader.__read_data__(train, self.tokenizer, max_len, flag = 'train'))
         self.val_data, self.text_val = ABSADataset(ABSADatasetReader.__read_data__(val, self.tokenizer, max_len, flag = 'val'))
         self.test_data, self.text_test = ABSADataset(ABSADatasetReader.__read_data__(test, self.tokenizer, max_len, flag = 'test'))
